+++ CHANGELOG.md
@@ -6,9 +6,11 @@
     - Dates & Times
     - If Statements & Comparison Operators
     - Switches
+    - Function Declarations & Expressions
 - An example with dates and times for Dates & Times was added.
 - An example with if statements and different operators for If Statements & Comparison Operators was added.
 - An example with switches was added.
+- An example with function declarations and expressions for Function Declarations & Expressions was added.
 
 ## August 4, 2020
 - Modern JavaScript From The Beginning (Udemy)